Extract VegaLite, Bokeh, Graphviz, Plotly charts
