+++ lib/streamlit/delta_generator.py
@@ -57,6 +57,10 @@ from streamlit.elements.doc_string import HelpMixin
 from streamlit.elements.exception_proto import ExceptionMixin
 from streamlit.elements.data_frame_proto import DataFrameMixin
 from streamlit.elements.altair import AltairMixin
+from streamlit.elements.bokeh_chart import BokehMixin
+from streamlit.elements.graphviz_chart import GraphvizMixin
+from streamlit.elements.plotly_chart import PlotlyMixin
+from streamlit.elements.vega_lite import VegaLiteMixin
 
 LOGGER = get_logger(__name__)
 
@@ -203,13 +207,17 @@ class DeltaGenerator(
     AlertMixin,
     AltairMixin,
     BalloonsMixin,
+    BokehMixin,
     ButtonMixin,
     DataFrameMixin,
     ExceptionMixin,
+    GraphvizMixin,
     HelpMixin,
     MarkdownMixin,
+    PlotlyMixin,
     JsonMixin,
     TextMixin,
+    VegaLiteMixin,
 ):
     """Creator of Delta protobuf messages.
 
@@ -482,302 +490,6 @@ class DeltaGenerator(
         return block_dg
 
     @_with_element
-    def vega_lite_chart(
-        self,
-        element,
-        data=None,
-        spec=None,
-        width=0,
-        use_container_width=False,
-        **kwargs,
-    ):
-        """Display a chart using the Vega-Lite library.
-
-        Parameters
-        ----------
-        data : pandas.DataFrame, pandas.Styler, numpy.ndarray, Iterable, dict,
-            or None
-            Either the data to be plotted or a Vega-Lite spec containing the
-            data (which more closely follows the Vega-Lite API).
-
-        spec : dict or None
-            The Vega-Lite spec for the chart. If the spec was already passed in
-            the previous argument, this must be set to None. See
-            https://vega.github.io/vega-lite/docs/ for more info.
-
-        width : number
-            Deprecated. If != 0 (default), will show an alert.
-            From now on you should set the width directly in the Vega-Lite
-            spec. Please refer to the Vega-Lite documentation for details.
-
-        use_container_width : bool
-            If True, set the chart width to the column width. This takes
-            precedence over Vega-Lite's native `width` value.
-
-        **kwargs : any
-            Same as spec, but as keywords.
-
-        Example
-        -------
-
-        >>> import pandas as pd
-        >>> import numpy as np
-        >>>
-        >>> df = pd.DataFrame(
-        ...     np.random.randn(200, 3),
-        ...     columns=['a', 'b', 'c'])
-        >>>
-        >>> st.vega_lite_chart(df, {
-        ...     'mark': {'type': 'circle', 'tooltip': True},
-        ...     'encoding': {
-        ...         'x': {'field': 'a', 'type': 'quantitative'},
-        ...         'y': {'field': 'b', 'type': 'quantitative'},
-        ...         'size': {'field': 'c', 'type': 'quantitative'},
-        ...         'color': {'field': 'c', 'type': 'quantitative'},
-        ...     },
-        ... })
-
-        .. output::
-           https://share.streamlit.io/0.25.0-2JkNY/index.html?id=8jmmXR8iKoZGV4kXaKGYV5
-           height: 200px
-
-        Examples of Vega-Lite usage without Streamlit can be found at
-        https://vega.github.io/vega-lite/examples/. Most of those can be easily
-        translated to the syntax shown above.
-
-        """
-        import streamlit.elements.vega_lite as vega_lite
-
-        if width != 0:
-            import streamlit as st
-
-            st.warning(
-                "The `width` argument in `st.vega_lite_chart` is deprecated and will be removed on 2020-03-04. To set the width, you should instead use Vega-Lite's native `width` argument as described at https://vega.github.io/vega-lite/docs/size.html"
-            )
-
-        vega_lite.marshall(
-            element.vega_lite_chart,
-            data,
-            spec,
-            use_container_width=use_container_width,
-            **kwargs,
-        )
-
-    @_with_element
-    def graphviz_chart(
-        self, element, figure_or_dot, width=0, height=0, use_container_width=False
-    ):
-        """Display a graph using the dagre-d3 library.
-
-        Parameters
-        ----------
-        figure_or_dot : graphviz.dot.Graph, graphviz.dot.Digraph, str
-            The Graphlib graph object or dot string to display
-
-        width : number
-            Deprecated. If != 0 (default), will show an alert.
-            From now on you should set the width directly in the Graphviz
-            spec. Please refer to the Graphviz documentation for details.
-
-        height : number
-            Deprecated. If != 0 (default), will show an alert.
-            From now on you should set the height directly in the Graphviz
-            spec. Please refer to the Graphviz documentation for details.
-
-        use_container_width : bool
-            If True, set the chart width to the column width. This takes
-            precedence over the figure's native `width` value.
-
-        Example
-        -------
-
-        >>> import streamlit as st
-        >>> import graphviz as graphviz
-        >>>
-        >>> # Create a graphlib graph object
-        >>> graph = graphviz.Digraph()
-        >>> graph.edge('run', 'intr')
-        >>> graph.edge('intr', 'runbl')
-        >>> graph.edge('runbl', 'run')
-        >>> graph.edge('run', 'kernel')
-        >>> graph.edge('kernel', 'zombie')
-        >>> graph.edge('kernel', 'sleep')
-        >>> graph.edge('kernel', 'runmem')
-        >>> graph.edge('sleep', 'swap')
-        >>> graph.edge('swap', 'runswap')
-        >>> graph.edge('runswap', 'new')
-        >>> graph.edge('runswap', 'runmem')
-        >>> graph.edge('new', 'runmem')
-        >>> graph.edge('sleep', 'runmem')
-        >>>
-        >>> st.graphviz_chart(graph)
-
-        Or you can render the chart from the graph using GraphViz's Dot
-        language:
-
-        >>> st.graphviz_chart('''
-            digraph {
-                run -> intr
-                intr -> runbl
-                runbl -> run
-                run -> kernel
-                kernel -> zombie
-                kernel -> sleep
-                kernel -> runmem
-                sleep -> swap
-                swap -> runswap
-                runswap -> new
-                runswap -> runmem
-                new -> runmem
-                sleep -> runmem
-            }
-        ''')
-
-        .. output::
-           https://share.streamlit.io/0.56.0-xTAd/index.html?id=GBn3GXZie5K1kXuBKe4yQL
-           height: 400px
-
-        """
-        import streamlit.elements.graphviz_chart as graphviz_chart
-
-        if width != 0 and height != 0:
-            import streamlit as st
-
-            st.warning(
-                "The `width` and `height` arguments in `st.graphviz` are deprecated and will be removed on 2020-03-04"
-            )
-        elif width != 0:
-            import streamlit as st
-
-            st.warning(
-                "The `width` argument in `st.graphviz` is deprecated and will be removed on 2020-03-04"
-            )
-        elif height != 0:
-            import streamlit as st
-
-            st.warning(
-                "The `height` argument in `st.graphviz` is deprecated and will be removed on 2020-03-04"
-            )
-
-        graphviz_chart.marshall(
-            element.graphviz_chart, figure_or_dot, use_container_width
-        )
-
-    @_with_element
-    def plotly_chart(
-        self,
-        element,
-        figure_or_data,
-        width=0,
-        height=0,
-        use_container_width=False,
-        sharing="streamlit",
-        **kwargs,
-    ):
-        """Display an interactive Plotly chart.
-
-        Plotly is a charting library for Python. The arguments to this function
-        closely follow the ones for Plotly's `plot()` function. You can find
-        more about Plotly at https://plot.ly/python.
-
-        Parameters
-        ----------
-        figure_or_data : plotly.graph_objs.Figure, plotly.graph_objs.Data,
-            dict/list of plotly.graph_objs.Figure/Data
-
-            See https://plot.ly/python/ for examples of graph descriptions.
-
-        width : int
-            Deprecated. If != 0 (default), will show an alert.
-            From now on you should set the width directly in the figure.
-            Please refer to the Plotly documentation for details.
-
-        height : int
-            Deprecated. If != 0 (default), will show an alert.
-            From now on you should set the height directly in the figure.
-            Please refer to the Plotly documentation for details.
-
-        use_container_width : bool
-            If True, set the chart width to the column width. This takes
-            precedence over the figure's native `width` value.
-
-        sharing : {'streamlit', 'private', 'secret', 'public'}
-            Use 'streamlit' to insert the plot and all its dependencies
-            directly in the Streamlit app, which means it works offline too.
-            This is the default.
-            Use any other sharing mode to send the app to Plotly's servers,
-            and embed the result into the Streamlit app. See
-            https://plot.ly/python/privacy/ for more. Note that these sharing
-            modes require a Plotly account.
-
-        **kwargs
-            Any argument accepted by Plotly's `plot()` function.
-
-
-        To show Plotly charts in Streamlit, just call `st.plotly_chart`
-        wherever you would call Plotly's `py.plot` or `py.iplot`.
-
-        Example
-        -------
-
-        The example below comes straight from the examples at
-        https://plot.ly/python:
-
-        >>> import streamlit as st
-        >>> import plotly.figure_factory as ff
-        >>> import numpy as np
-        >>>
-        >>> # Add histogram data
-        >>> x1 = np.random.randn(200) - 2
-        >>> x2 = np.random.randn(200)
-        >>> x3 = np.random.randn(200) + 2
-        >>>
-        >>> # Group data together
-        >>> hist_data = [x1, x2, x3]
-        >>>
-        >>> group_labels = ['Group 1', 'Group 2', 'Group 3']
-        >>>
-        >>> # Create distplot with custom bin_size
-        >>> fig = ff.create_distplot(
-        ...         hist_data, group_labels, bin_size=[.1, .25, .5])
-        >>>
-        >>> # Plot!
-        >>> st.plotly_chart(fig, use_container_width=True)
-
-        .. output::
-           https://share.streamlit.io/0.56.0-xTAd/index.html?id=TuP96xX8JnsoQeUGAPjkGQ
-           height: 400px
-
-        """
-        # NOTE: "figure_or_data" is the name used in Plotly's .plot() method
-        # for their main parameter. I don't like the name, but it's best to
-        # keep it in sync with what Plotly calls it.
-        import streamlit.elements.plotly_chart as plotly_chart
-
-        if width != 0 and height != 0:
-            import streamlit as st
-
-            st.warning(
-                "The `width` and `height` arguments in `st.plotly_chart` are deprecated and will be removed on 2020-03-04. To set these values, you should instead use Plotly's native arguments as described at https://plot.ly/python/setting-graph-size/"
-            )
-        elif width != 0:
-            import streamlit as st
-
-            st.warning(
-                "The `width` argument in `st.plotly_chart` is deprecated and will be removed on 2020-03-04. To set the width, you should instead use Plotly's native `width` argument as described at https://plot.ly/python/setting-graph-size/"
-            )
-        elif height != 0:
-            import streamlit as st
-
-            st.warning(
-                "The `height` argument in `st.plotly_chart` is deprecated and will be removed on 2020-03-04. To set the height, you should instead use Plotly's native `height` argument as described at https://plot.ly/python/setting-graph-size/"
-            )
-
-        plotly_chart.marshall(
-            element.plotly_chart, figure_or_data, use_container_width, sharing, **kwargs
-        )
-
-    @_with_element
     def pyplot(self, element, fig=None, clear_figure=None, **kwargs):
         """Display a matplotlib.pyplot figure.
 
@@ -828,52 +540,6 @@ class DeltaGenerator(
         import streamlit.elements.pyplot as pyplot
 
         pyplot.marshall(self._get_coordinates, element, fig, clear_figure, **kwargs)
-
-    @_with_element
-    def bokeh_chart(self, element, figure, use_container_width=False):
-        """Display an interactive Bokeh chart.
-
-        Bokeh is a charting library for Python. The arguments to this function
-        closely follow the ones for Bokeh's `show` function. You can find
-        more about Bokeh at https://bokeh.pydata.org.
-
-        Parameters
-        ----------
-        figure : bokeh.plotting.figure.Figure
-            A Bokeh figure to plot.
-
-        use_container_width : bool
-            If True, set the chart width to the column width. This takes
-            precedence over Bokeh's native `width` value.
-
-        To show Bokeh charts in Streamlit, just call `st.bokeh_chart`
-        wherever you would call Bokeh's `show`.
-
-        Example
-        -------
-        >>> import streamlit as st
-        >>> from bokeh.plotting import figure
-        >>>
-        >>> x = [1, 2, 3, 4, 5]
-        >>> y = [6, 7, 2, 4, 5]
-        >>>
-        >>> p = figure(
-        ...     title='simple line example',
-        ...     x_axis_label='x',
-        ...     y_axis_label='y')
-        ...
-        >>> p.line(x, y, legend='Trend', line_width=2)
-        >>>
-        >>> st.bokeh_chart(p, use_container_width=True)
-
-        .. output::
-           https://share.streamlit.io/0.56.0-xTAd/index.html?id=Fdhg51uMbGMLRRxXV6ubzp
-           height: 600px
-
-        """
-        import streamlit.elements.bokeh_chart as bokeh_chart
-
-        bokeh_chart.marshall(element.bokeh_chart, figure, use_container_width)
 
     @_with_element
     def image(
